misc(cassandra): Stop querying the chunk index table and use the regular table instead.
