+++ cassandra/src/main/scala/filodb.cassandra/columnstore/IndexTable.scala
@@ -4,20 +4,15 @@ import java.nio.ByteBuffer
 
 import scala.concurrent.{ExecutionContext, Future}
 
-import com.datastax.driver.core.{ConsistencyLevel, Row}
-import monix.reactive.Observable
+import com.datastax.driver.core.ConsistencyLevel
 
 import filodb.cassandra.FiloCassandraConnector
 import filodb.core._
 import filodb.core.store.ChunkSinkStats
-import filodb.memory.format.UnsafeUtils
-
-// Typical record read from serialized incremental index (ChunkInfo + Skips) entries
-case class IndexRecord(binPartition: ByteBuffer, data: ByteBuffer) {
-  def partBaseOffset: (Any, Long, Int) = UnsafeUtils.BOLfromBuffer(binPartition)
-}
 
 /**
+ * FIXME: To be removed. The table is populated but never queried.
+ *
  * Represents the table which holds the incremental chunk metadata or indexes for each segment
  * of a partition.  The chunk index contains info about each chunk including if it overrides
  * any records in previous chunks.  Each new chunkSet that is written results in a tiny bit more metadata.
@@ -31,7 +26,6 @@ case class IndexRecord(binPartition: ByteBuffer, data: ByteBuffer) {
  */
 sealed class IndexTable(val dataset: DatasetRef, val connector: FiloCassandraConnector)
                        (implicit ec: ExecutionContext) extends BaseDatasetTable {
-  import scala.collection.JavaConverters._
 
   import filodb.cassandra.Util._
 
@@ -45,44 +39,6 @@ sealed class IndexTable(val dataset: DatasetRef, val connector: FiloCassandraCon
                      |    PRIMARY KEY (partition, indextype, chunkid)
                      |) WITH compression = {
                     'sstable_compression': '$sstableCompression'}""".stripMargin
-
-
-  def fromRow(row: Row): IndexRecord =
-    IndexRecord(row.getBytes("partition"), row.getBytes("data"))
-
-  val selectCql = s"SELECT partition, data FROM $tableString WHERE "
-  val partitionFilter = "partition = ? AND indextype = 1"
-  val partInFilter = "partition IN ? AND indextype = 1"
-  lazy val allPartReadCql = session.prepare(selectCql + partitionFilter)
-  lazy val inPartReadCql = session.prepare(selectCql + partInFilter)
-
-  /**
-   * Retrieves all indices from a single partition.
-   */
-  def getIndices(binPartition: Array[Byte]): Observable[IndexRecord] = {
-    val it = session.execute(allPartReadCql.bind(toBuffer(binPartition)))
-                    .asScala.toIterator.map(fromRow)
-    Observable.fromIterator(it).handleObservableErrors
-  }
-
-  def getMultiIndices(partitions: Seq[Array[Byte]]): Observable[IndexRecord] = {
-    val query = inPartReadCql.bind().setList(0, partitions.map(toBuffer).asJava, classOf[ByteBuffer])
-    val it = session.execute(query).asScala.toIterator.map(fromRow)
-    Observable.fromIterator(it).handleObservableErrors
-  }
-
-  val tokenQ = "TOKEN(partition)"
-
-  def scanIndices(tokens: Seq[(String, String)]): Observable[IndexRecord] = {
-    def cql(start: String, end: String): String =
-      s"SELECT * FROM $tableString WHERE $tokenQ >= $start AND $tokenQ < $end AND indextype = 1 " +
-      s"ALLOW FILTERING"
-    val it = tokens.iterator.flatMap { case (start, end) =>
-        session.execute(cql(start, end)).iterator.asScala
-               .map { row => fromRow(row) }
-      }
-    Observable.fromIterator(it).handleObservableErrors
-  }
 
   lazy val writeIndexCql = session.prepare(
     s"INSERT INTO $tableString (partition, indextype, chunkid, data) " +
